# chore: introduce an optimized integer compare algorithm for lists